♻️ move platform detect into matrix utils
