+++ app/lib/matrix-api/utils.ts
@@ -3,6 +3,7 @@ import type { Room } from "matrix-js-sdk";
 import { EventTimeline } from "matrix-js-sdk/src/models/event-timeline";
 import * as sdk from "matrix-js-sdk";
 import { client } from "./client";
+import type { PlatformEnum } from "~/lib/contacts-server-api/types";
 
 /**
  * Get base URL from user ID
@@ -224,4 +225,30 @@ export function getImageBlob(
 
   imageCache.set(mxcUrl, promise);
   return promise;
+}
+
+/**
+ * Detect platform from bridge information in a Matrix room
+ */
+export function detectPlatform(room: Room): PlatformEnum {
+  const bridgeStateEvents = room
+    .getLiveTimeline()
+    .getState(EventTimeline.FORWARDS)
+    ?.getStateEvents("m.bridge");
+
+  if (!bridgeStateEvents || bridgeStateEvents.length === 0) {
+    return "Matrix" as PlatformEnum;
+  }
+
+  const content = bridgeStateEvents[0].getContent();
+  const protocol = content?.protocol?.id;
+
+  switch (protocol) {
+    case "discord":
+      return "Discord" as PlatformEnum;
+    case "telegram":
+      return "Telegram" as PlatformEnum;
+    default:
+      return "Matrix" as PlatformEnum;
+  }
 }
